modify login function.
