+++ lib/databases/databases.dart
@@ -1,0 +1,1 @@
+export 'scores/user_score.dart';
