+++ lib/databases/scores/user_scores.dart
@@ -1,0 +1,1 @@
+export 'user_score.dart';
