+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -6,7 +7,6 @@ import 'package:nibuichi/screens/games/game.dart';
 import 'package:nibuichi/screens/hubs/hubs.dart';
 import 'package:nibuichi/screens/hubs/hubScreen.dart';
 import 'package:nibuichi/screens/logins/logins.dart';
-
 import 'firebase_options.dart';
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -47,6 +47,7 @@ Future<void> main() async {
     await Firebase.initializeApp(
       options: DefaultFirebaseOptions.currentPlatform,
     );
+    FirebaseDatabase.instance.setPersistenceEnabled(true);
 
   runApp(
       const ProviderScope(
