+++ lib/providers/login_provider.dart
@@ -2,6 +2,6 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 ////////////////////////////////////////////////////////////////////////////////
 
-final hubIndexProvider = StateProvider<int>((ref)=>0);
+final keyProvider = StateProvider((ref) => "splash");
 
 ////////////////////////////////////////////////////////////////////////////////
