+++ lib/providers/ranking_provider.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nibuichi/databases/databases.dart';
+
+final scoreProvider = StateProvider((ref) => UserScore(highScore: 0));
