+++ lib/screens/logins/common_data.dart
@@ -1,0 +1,2 @@
+const double n = 8.5;
+const double padding  = 32;
