+++ pubspec.yaml
@@ -21,6 +21,11 @@ dependencies:
   hooks_riverpod: ^2.5.1
   uuid: ^4.4.2
   cloud_firestore: ^5.2.1
+  freezed: ^2.4.4
+  freezed_annotation: ^2.4.4
+  build_runner: ^2.4.11
+  json_annotation: ^4.9.0
+  json_serializable: ^6.8.0
 
 dev_dependencies:
   flutter_test:
